fix a typo

+++ scripts/plot_quality_results.cpp
@@ -139,7 +139,7 @@ void plot_quality_results(const char* path = "../Rootfiles/cluster_0_fermilab_be
     {
         double eff = 0;
         if(m_layer_didhits_tracker_based.find(i.first) != m_layer_didhits_tracker_based.end())
-            eff = m_layer_didhits_tracker_based[i.first]/i.second;
+            eff = (double)m_layer_didhits_tracker_based[i.first]/(double)i.second;
         string tracker_based_eff = Form("layer %d, tracker based efficiency = %.2f", i.first, eff);
         latex.DrawLatex(0.1, 0.85 - .05*_c, tracker_based_eff.c_str());
         _c++;
